# references/openapi.yaml
openapi: 3.1.0
info:
  title: DataKB API
  version: 1.0.0
  description: |
    REST API for DataKB — an open source executable knowledge graph for data
    engineering teams. Notebooks are attached to graph nodes and executed in
    isolated per-user Jupyter kernel sessions. GCP credentials are injected
    at kernel startup via server-side Service Account impersonation.

    ## Authentication

    All endpoints except `/api/auth/login`, `/api/auth/google`,
    `/api/auth/callback`, and `/api/health` require a valid JWT:

    ```
    Authorization: Bearer <jwt>
    ```

    JWTs are issued on login (local mode) or after Google OAuth (google mode).
    They expire after 8 hours. Use `POST /api/auth/refresh` with the
    `datakb_refresh` httpOnly cookie to obtain a new JWT silently.

    ## Roles

    | Role    | Capabilities |
    |---------|-------------|
    | `admin`  | Full access — graph, notebooks, kernels, GCP config, users |
    | `editor` | Graph read/write, notebooks read/write, kernels start/stop |
    | `viewer` | Graph read, notebook outputs read (static, no execution) |

    ## Error format

    All error responses use a consistent shape:

    ```json
    {
      "error": "error_code",
      "message": "Human-readable description",
      "details": {}
    }
    ```

  contact:
    name: DataKB on GitHub
    url: https://github.com/datakb/datakb
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    ISODateTime:
      type: string
      format: date-time
      example: "2026-04-17T14:32:11Z"
    NodeType:
      type: string
      enum: [service, pipeline, database, redis, gcs_bucket, schema, note, team]
      example: redis
    EdgeLabel:
      type: string
      enum: [reads_from, writes_to, depends_on, owns, produces, consumes, triggers]
      example: reads_from
    UserRole:
      type: string
      enum: [admin, editor, viewer]
      example: editor
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: not_found
        message:
          type: string
          example: "Node with id abc123 not found"
        details:
          type: object
          additionalProperties: true

    # ── Auth ────────────────────────────────────────────────────
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: alice@company.com
        password:
          type: string
          format: password
          minLength: 8
    TokenResponse:
      type: object
      required: [access_token, token_type, expires_in, user]
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          enum: [bearer]
        expires_in:
          type: integer
          description: Seconds until expiry
          example: 28800
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, role, created_at]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        email:
          type: string
          format: email
          example: alice@company.com
        display_name:
          type: string
          nullable: true
          example: Alice
        role:
          $ref: "#/components/schemas/UserRole"
        team:
          type: string
          nullable: true
          example: platform
        is_active:
          type: boolean
          example: true
        created_at:
          $ref: "#/components/schemas/ISODateTime"
        last_login_at:
          $ref: "#/components/schemas/ISODateTime"
          nullable: true
...
    GCPProjectSummary:
      type: object
      required: [id, project_id, display_name]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        project_id:
          type: string
          example: payments-prod
        display_name:
          type: string
          example: "Payments Production"
    ServiceAccountSummary:
      type: object
      required: [id, sa_email, display_name]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        sa_email:
          type: string
          example: platform-sa@payments-prod.iam.gserviceaccount.com
        display_name:
          type: string
          example: "Platform SA"
    Node:
      type: object
      required:
        - id
        - title
        - type
        - tags
        - resource_bindings
        - created_by
        - created_at
        - updated_at
        - is_archived
        - current_user_access
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        title:
          type: string
          example: "Auth Service Redis Cache"
        type:
          $ref: "#/components/schemas/NodeType"
        team:
          type: string
          nullable: true
          example: platform
        description:
          type: string
          nullable: true
          example: "Session token cache. TTL 24h. Eviction: allkeys-lru."
        tags:
          type: array
          items:
            type: string
          example: [redis, cache, auth]
        notebook_path:
          type: string
          nullable: true
          description: Relative path within the storage backend
          example: "platform/auth-redis.ipynb"
        gcp_project:
          $ref: "#/components/schemas/GCPProjectSummary"
          nullable: true
        service_account:
          $ref: "#/components/schemas/ServiceAccountSummary"
          nullable: true
        resource_bindings:
          type: array
          items:
            type: string
          description: Names of resources injected into this node's kernels
          example: [payments-redis, payments-bq]
        created_by:
          type: string
          format: email
          example: alice@company.com
        created_at:
          $ref: "#/components/schemas/ISODateTime"
        updated_at:
          $ref: "#/components/schemas/ISODateTime"
        last_run_at:
          $ref: "#/components/schemas/ISODateTime"
          nullable: true
        last_run_by:
          type: string
          format: email
          nullable: true
          example: bob@company.com
        active_kernel_count:
          type: integer
          description: Currently active kernel sessions across all users
          example: 2
        is_archived:
          type: boolean
          example: false
        current_user_access:
          type: string
          enum: [none, view, run, edit]
          description: Resolved access level for the requesting user
          example: run
    Edge:
      type: object
      required: [id, source_id, target_id, label, created_by, created_at]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        source_id:
          $ref: "#/components/schemas/UUID"
        target_id:
          $ref: "#/components/schemas/UUID"
        label:
          $ref: "#/components/schemas/EdgeLabel"
        created_by:
          type: string
          format: email
          example: alice@company.com
        created_at:
          $ref: "#/components/schemas/ISODateTime"
    GraphResponse:
      type: object
      required: [nodes, edges]
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/Node"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/Edge"
    CreateNodeRequest:
      type: object
      required: [title, type]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: "Auth Service Redis Cache"
        type:
          $ref: "#/components/schemas/NodeType"
        team:
          type: string
          nullable: true
          example: platform
        description:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
          default: []
          example: [redis, cache, auth]
        notebook_path:
          type: string
          nullable: true
          example: "platform/auth-redis.ipynb"
        gcp_project_id:
          $ref: "#/components/schemas/UUID"
          nullable: true
        sa_id:
          $ref: "#/components/schemas/UUID"
          nullable: true
        resource_bindings:
          type: array
          items:
            type: string
          default: []
          example: [payments-redis]
    UpdateNodeRequest:
      type: object
      description: All fields optional — only provided fields are updated
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        team:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        notebook_path:
          type: string
          nullable: true
        gcp_project_id:
          $ref: "#/components/schemas/UUID"
          nullable: true
        sa_id:
          $ref: "#/components/schemas/UUID"
          nullable: true
        resource_bindings:
          type: array
          items:
            type: string
        is_archived:
          type: boolean
    CreateEdgeRequest:
      type: object
      required: [source_id, target_id, label]
      properties:
        source_id:
          $ref: "#/components/schemas/UUID"
        target_id:
          $ref: "#/components/schemas/UUID"
        label:
          $ref: "#/components/schemas/EdgeLabel"

    # ── Notebooks ───────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: unauthorized
            message: "Missing or invalid authorization token."
    Forbidden:
      description: Authenticated but insufficient role
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: forbidden
            message: "Your role does not permit this action."
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: not_found
            message: "The requested resource was not found."
...
    UnprocessableEntity:
      description: Request body validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: validation_error
            message: "Request body is invalid."
            details:
              field: title
              issue: "Field is required."
...
paths:
...
  /api/auth/login:
    post:
      tags: [auth]
      summary: Local login
      description: |
        Authenticate with email + password. Returns a JWT and sets the
        `datakb_refresh` httpOnly cookie. Only available when AUTH_MODE=local.
      security: []
      operationId: auth_login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
                example: "datakb_refresh=<token>; HttpOnly; SameSite=Strict; Max-Age=2592000"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: invalid_credentials
                message: "Email or password is incorrect."
        "403":
          description: AUTH_MODE is not local
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: auth_mode_mismatch
                message: "This instance uses Google OAuth. Use /api/auth/google."
...
  /api/auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token
      description: |
        Issues a new access token using the `datakb_refresh` httpOnly cookie.
        Called automatically by the frontend on 401 responses.
      security: []
      operationId: auth_refresh
      responses:
        "200":
          description: New token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Refresh token missing, invalid, or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: refresh_token_expired
                message: "Refresh token has expired. Please log in again."
  /api/auth/me:
    get:
      tags: [auth]
      summary: Get current user
      operationId: auth_me
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ════════════════════════════════════════
  # GRAPH
  # ════════════════════════════════════════
  /api/graph:
    get:
      tags: [graph]
      summary: Get full graph
      description: |
        Returns all non-archived nodes and all edges.
        The `current_user_access` field on each node reflects the
        requesting user's resolved access level.
      operationId: graph_get
      parameters:
        - name: include_archived
          in: query
          schema:
            type: boolean
            default: false
        - name: team
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: array
            items:
              $ref: "#/components/schemas/NodeType"
          style: form
          explode: true
      responses:
        "200":
          description: Graph nodes and edges
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/nodes:
    post:
      tags: [graph]
      summary: Create node
      operationId: nodes_create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNodeRequest"
      responses:
        "201":
          description: Node created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: A node with this title already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: duplicate_title
                message: "A node with title 'Auth Service Redis Cache' already exists."
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/nodes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"

    get:
      tags: [graph]
      summary: Get node
      operationId: nodes_get
      responses:
        "200":
          description: Node
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [graph]
      summary: Update node
      description: Partial update — only provided fields are changed.
      operationId: nodes_update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNodeRequest"
      responses:
        "200":
          description: Updated node
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"

    delete:
      tags: [graph]
      summary: Archive node
      description: |
        Soft-deletes by setting `is_archived = true`. The notebook file
        in storage is NOT deleted. Restore via PUT with `is_archived: false`.
      operationId: nodes_archive
      responses:
        "204":
          description: Node archived
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/edges:
    post:
      tags: [graph]
      summary: Create edge
      description: The (source_id, target_id, label) triple must be unique.
      operationId: edges_create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEdgeRequest"
      responses:
        "201":
          description: Edge created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Edge"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Source or target node not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Edge already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: duplicate_edge
                message: "An edge with this source, target, and label already exists."
  /api/edges/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"

    delete:
      tags: [graph]
      summary: Delete edge
      operationId: edges_delete
      responses:
        "204":
          description: Edge deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ════════════════════════════════════════
  # NOTEBOOKS
  # ════════════════════════════════════════
...